requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0

# Document processing
pyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
pandas>=2.0.0

# NLP and LLM integration
langchain>=0.0.267
openai>=0.27.8
huggingface-hub>=0.16.4
transformers>=4.30.2
sentence-transformers>=2.2.2

# Vector database
chromadb>=0.4.6

# Testing
pytest>=7.3.1

# Development
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0